% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data Based on Configuration}
\usage{
clean_data(sheet_data, config)
}
\arguments{
\item{sheet_data}{A data frame to clean.}

\item{config}{A list of cleaning options.}
}
\value{
A cleaned data frame.
}
\description{
Applies cleaning steps to a data frame based on the provided configuration.
}
