% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_config.R
\name{load_config}
\alias{load_config}
\title{Load Configuration}
\usage{
load_config(file)
}
\arguments{
\item{file}{Path to the spreadsheet file.}
}
\value{
A list containing the configuration options.
}
\description{
Loads the appropriate configuration file for a given spreadsheet.
First checks for a spreadsheet-specific configuration file in the current working directory,
and falls back to the default configuration in the user's home directory.
}
\examples{
config <- load_config(".excel2cleantibr_config.json")
}
