% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_tibbles.R
\name{read_excel_tibbles}
\alias{read_excel_tibbles}
\title{Read Excel File and Import Sheets as Tibbles}
\usage{
read_excel_tibbles(file, config = NULL, output_format = "list")
}
\arguments{
\item{file}{Path to the Excel file.}

\item{config}{Path to a configuration file (optional). If NULL, the function attempts to load
a configuration file using `load_config()`.}

\item{output_format}{Output format: "list" (default) or "global_env".}
}
\value{
A named list of tibbles or NULL (if output to the global environment).
}
\description{
Reads an Excel file and converts its sheets into tibbles, applying configurable cleaning and naming conventions.
}
\examples{
read_excel_tibbles("example.xlsx")
}
