% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_sheet_names.R
\name{sanitize_sheet_names}
\alias{sanitize_sheet_names}
\title{Sanitize Sheet Names}
\usage{
sanitize_sheet_names(names, custom_function = NULL)
}
\arguments{
\item{names}{A character vector of sheet names.}

\item{custom_function}{A user-defined function for sanitization (optional).}
}
\value{
Sanitized sheet names.
}
\description{
Transforms sheet names to meet naming conventions.
}
