% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_default_config.R
\name{save_default_config}
\alias{save_default_config}
\title{Save Default Configuration}
\usage{
save_default_config()
}
\value{
A message indicating the file was saved.
}
\description{
Saves a default configuration file to the user's home directory for use by `read_excel_tibbles`.
}
\examples{
save_default_config()
}
